Moved BrowserWindow and its abstraction interfaces to other source files
